Issue #2533: Makefile cleanup

+++ Makefile
@@ -1,6 +1,6 @@
 install:
 	if [ ! -f docker-compose.local.yml ]; then cp docker-compose.local.example.yml docker-compose.local.yml; fi
-	-docker-compose -f docker-compose.yml -f docker-compose.local.yml down
+	if [[ $$(docker-compose ps -q) ]]; then docker-compose -f docker-compose.yml -f docker-compose.local.yml down; fi
 	docker-compose -f docker-compose.yml -f docker-compose.local.yml up -d --build --force-recreate
 	make reset_db
 	make reset_logs
@@ -12,10 +12,8 @@ build:
 	docker build -t savaslabs/squabble .
 
 reset_logs:
-	-docker exec squabble_app_1 rm /var/www/html/storage/logs/lumen.log
+	docker exec squabble_app_1 sh -c "if [ -f /var/www/html/storage/logs/lumen.log]; then rm /var/www/html/storage/logs/lumen.log; fi"
 	docker exec squabble_app_1 touch /var/www/html/storage/logs/lumen.log
-	-docker exec squabble_app_1 chmod a+w /var/www/html/storage/logs
-	-docker exec squabble_app_1 chmod a+w /var/www/html/storage/logs/lumen.log
 
 logs:
 	docker logs -f squabble_app_1
@@ -26,19 +24,17 @@ clean:
 	-docker volume rm squabble_db
 
 reset_db:
-	-docker exec squabble_app_1 rm /db/database.sqlite
+	docker exec squabble_app_1 sh -c "if [ -f /db/database.sqlite ]; then rm /db/database.sqlite; fi"
 	docker exec squabble_app_1 touch /db/database.sqlite
-	-docker exec squabble_app_1 chmod a+w /db/database.sqlite
-	-docker exec squabble_app_1 chmod a+w /db
-	-docker exec squabble_app_1 php artisan migrate
+	docker exec squabble_app_1 php artisan migrate
 
 up:
-	- docker-compose -f docker-compose.yml -f docker-compose.local.yml up -d
-	- docker-compose ps
+	docker-compose -f docker-compose.yml -f docker-compose.local.yml up -d
+	docker-compose ps
 
 down:
-	- docker-compose -f docker-compose.yml -f docker-compose.local.yml down
-	- docker-compose ps
+	docker-compose -f docker-compose.yml -f docker-compose.local.yml down
+	docker-compose ps
 
 phpunit:
 	make reset_db
@@ -51,7 +47,6 @@ behat:
 wip:
 	make reset_db
 	docker-compose -f behat/docker-compose.yml run --rm behat --tags=@wip
-
 
 test:
 	make phpunit
